Move recipe into hilt subpackage
