+++ app/src/main/AndroidManifest.xml
@@ -86,7 +86,7 @@
             android:exported="true"
             android:theme="@style/Theme.Nav3Recipes"/>
         <activity
-            android:name=".modular.ModularActivity"
+            android:name=".modular.hilt.ModularActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Nav3Recipes"/>
