+++ app/src/main/java/com/example/nav3recipes/modular/hilt/CommonModule.kt
@@ -1,12 +1,8 @@
-package com.example.nav3recipes.modular
+package com.example.nav3recipes.modular.hilt
 
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.navigation3.runtime.EntryProviderBuilder
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityRetainedComponent
 import dagger.hilt.android.scopes.ActivityRetainedScoped
 import javax.inject.Inject
