+++ app/src/main/java/com/example/nav3recipes/modular/hilt/ConversationModule.kt
@@ -1,4 +1,4 @@
-package com.example.nav3recipes.modular
+package com.example.nav3recipes.modular.hilt
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -16,7 +16,6 @@ import androidx.compose.material3.ListItemDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
