+++ app/src/main/java/com/example/nav3recipes/modular/hilt/ModularActivity.kt
@@ -1,14 +1,12 @@
-package com.example.nav3recipes.modular
+package com.example.nav3recipes.modular.hilt
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
-import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Modifier
 import androidx.navigation3.ui.NavDisplay
-import androidx.navigation3.runtime.EntryProviderBuilder
 import androidx.navigation3.runtime.entryProvider
 import com.example.nav3recipes.ui.setEdgeToEdgeConfig
 import dagger.hilt.android.AndroidEntryPoint
@@ -16,7 +14,8 @@ import javax.inject.Inject
 
 /**
  * This recipe demonstrates how to use a modular approach with Navigation 3,
- * where different parts of the application are defined in separate modules.
+ * where different parts of the application are defined in separate modules and injected
+ * into the main app using Dagger/Hilt.
  * 
  * Features (Conversation and Profile) are split into two modules: 
  * - api: defines the public facing routes for this feature
